+++ irs-api/src/test/java/org/eclipse/tractusx/irs/IrsFunctionalTest.java
@@ -88,7 +88,7 @@ class IrsFunctionalTest {
                                                      .until(getJobDetails(jobHandle),
                                                              jobs -> jobs.isPresent() && jobs.get()
                                                                                              .getJob()
-                                                                                             .getJobState()
+                                                                                             .getState()
                                                                                              .equals(JobState.COMPLETED));
 
         assertThat(finishedJob).isPresent();